Remove Elasticsearch admin commands
