+++ tools/cli/admin.go
@@ -403,66 +403,6 @@ func newAdminNamespaceCommands() []cli.Command {
 	}
 }
 
-func newAdminElasticSearchCommands() []cli.Command {
-	return []cli.Command{
-		{
-			Name:    "catIndex",
-			Aliases: []string{"cind"},
-			Usage:   "Cat Indices on Elasticsearch",
-			Flags:   getESFlags(false),
-			Action: func(c *cli.Context) {
-				AdminCatIndices(c)
-			},
-		},
-		{
-			Name:    "index",
-			Aliases: []string{"ind"},
-			Usage:   "Index docs on Elasticsearch",
-			Flags: append(
-				getESFlags(true),
-				cli.StringFlag{
-					Name:  FlagInputFileWithAlias,
-					Usage: "Input file of indexerspb.Message in json format, separated by newline",
-				},
-				cli.IntFlag{
-					Name:  FlagBatchSizeWithAlias,
-					Usage: "Optional batch size of actions for bulk operations",
-					Value: 10,
-				},
-			),
-			Action: func(c *cli.Context) {
-				AdminIndex(c)
-			},
-		},
-		{
-			Name:    "delete",
-			Aliases: []string{"del"},
-			Usage:   "Delete docs on Elasticsearch",
-			Flags: append(
-				getESFlags(true),
-				cli.StringFlag{
-					Name: FlagInputFileWithAlias,
-					Usage: "Input file name. Redirect temporal wf list result (with table format) to a file and use as delete input. " +
-						"First line should be table header like WORKFLOW TYPE | WORKFLOW ID | RUN ID | ...",
-				},
-				cli.IntFlag{
-					Name:  FlagBatchSizeWithAlias,
-					Usage: "Optional batch size of actions for bulk operations",
-					Value: 1000,
-				},
-				cli.IntFlag{
-					Name:  FlagRPS,
-					Usage: "Optional batch request rate per second",
-					Value: 30,
-				},
-			),
-			Action: func(c *cli.Context) {
-				AdminDelete(c)
-			},
-		},
-	}
-}
-
 func newAdminTaskQueueCommands() []cli.Command {
 	return []cli.Command{
 		{
