+++ tools/cli/flags.go
@@ -90,14 +90,7 @@ const (
 	FlagSkipBaseIsNotCurrent                  = "skip_base_is_not_current"
 	FlagDryRun                                = "dry_run"
 	FlagNonDeterministicOnly                  = "only_non_deterministic"
-	FlagInputTopic                            = "input_topic"
-	FlagInputTopicWithAlias                   = FlagInputTopic + ", it"
-	FlagHostFile                              = "host_file"
 	FlagCluster                               = "cluster"
-	FlagInputCluster                          = "input_cluster"
-	FlagStartOffset                           = "start_offset"
-	FlagTopic                                 = "topic"
-	FlagGroup                                 = "group"
 	FlagResult                                = "result"
 	FlagIdentity                              = "identity"
 	FlagDetail                                = "detail"
@@ -176,20 +169,13 @@ const (
 	FlagMaxFieldLengthWithAlias               = FlagMaxFieldLength + ", maxl"
 	FlagSkipErrorMode                         = "skip_errors"
 	FlagSkipErrorModeWithAlias                = FlagSkipErrorMode + ", serr"
-	FlagHeadersMode                           = "headers"
-	FlagHeadersModeWithAlias                  = FlagHeadersMode + ", he"
-	FlagMessageType                           = "message_type"
-	FlagMessageTypeWithAlias                  = FlagMessageType + ", mt"
 	FlagURL                                   = "url"
 	FlagIndex                                 = "index"
-	FlagBatchSize                             = "batch_size"
-	FlagBatchSizeWithAlias                    = FlagBatchSize + ", bs"
 	FlagMemoKey                               = "memo_key"
 	FlagMemo                                  = "memo"
 	FlagMemoFile                              = "memo_file"
 	FlagSearchAttributeKey                    = "search_attr_key"
 	FlagSearchAttributeValue                  = "search_attr_value"
-	FlagSearchAttributeType                   = "search_attr_type"
 	FlagAddBadBinary                          = "add_bad_binary"
 	FlagRemoveBadBinary                       = "remove_bad_binary"
 	FlagResetType                             = "reset_type"
